Changed height type to size_t, reason :
Map can contain more than MAX_INT, so height can overflow if so

+++ utils/tree.hpp
@@ -13,7 +13,7 @@ namespace ft
                 avlTree *right;
                 avlTree *left;
                 avlTree *parent;
-                int height;
+                size_t height;
 
             public :
                 avlTree( ) { }
